add applications of TopTrader
